Add product price schedulers

+++ src/main/java/tk/project/goodsstorage/product/dto/ProductDto.java
@@ -24,7 +24,7 @@ public class ProductDto {
 
     String category;
 
-    Integer price;
+    Double price;
 
     Long count;
